+++ src/components/Header/Header.css
@@ -1,11 +1,11 @@
 .nav {
   /* flex box layout */
   display: flex;
-  justify-content: space-between;
+  /* justify-content: space-between; */
   align-items: center;
 
   margin: 0 0 30px;
-  background-color: #00acb0;
+  background-color: #4c4000;
   overflow: hidden;
 }
 
@@ -16,15 +16,15 @@
 .nav-title {
   font-size: 24px;
   font-weight: 700;
-  color: #f2f2f2;
+  color: white;
   padding-left: 10px;
   margin: 0;
 }
 
 .navLink {
   display: inline-block;
-  color: #f2f2f2;
-  background-color: #00acb0;
+  color: white;
+  background-color: #4c4000;
   text-align: center;
   padding: 24px 10px;
   text-decoration: none;
@@ -38,5 +38,5 @@
 }
 
 .navLink:hover {
-  background-color: #008183;
+  background-color: #6f6633;
 }
